Moved function to service file
